refactor(suggest): pull up top token generation to suggestion ctx...
...from TermConceptSuggester
